+++ internal/models/roles_models.go
@@ -3,7 +3,7 @@ package models
 import "time"
 
 type Role struct {
-	ID              uint             `gorm:"primaryKey;autoIncrement" json:"id"`
+	ID              int64            `gorm:"primaryKey;autoIncrement" json:"id"`
 	Name            string           `gorm:"unique;not null;size:75" json:"name"`
 	IsGlobal        bool             `gorm:"not null" json:"is_global"`
 	Description     string           `gorm:"size:255" json:"description"`
@@ -14,7 +14,7 @@ type Role struct {
 }
 
 type Permission struct {
-	ID              uint             `gorm:"primaryKey;autoIncrement" json:"id"`
+	ID              int64            `gorm:"primaryKey;autoIncrement" json:"id"`
 	Name            string           `gorm:"unique;not null;size:75" json:"name"`
 	Description     string           `gorm:"size:255" json:"description"`
 	CreatedAt       time.Time        `json:"created_at"`
@@ -23,9 +23,9 @@ type Permission struct {
 }
 
 type RolePermission struct {
-	ID           uint       `gorm:"primaryKey;autoIncrement" json:"id"`
-	RoleID       uint       `gorm:"not null;uniqueIndex:idx_role_permission" json:"role_id"`
-	PermissionID uint       `gorm:"not null;uniqueIndex:idx_role_permission" json:"permission_id"`
+	ID           int64      `gorm:"primaryKey;autoIncrement" json:"id"`
+	RoleID       int64      `gorm:"not null;uniqueIndex:idx_role_permission" json:"role_id"`
+	PermissionID int64      `gorm:"not null;uniqueIndex:idx_role_permission" json:"permission_id"`
 	Role         Role       `gorm:"constraint:OnDelete:CASCADE;foreignKey:RoleID" json:"-"`
 	Permission   Permission `gorm:"constraint:OnDelete:CASCADE;foreignKey:PermissionID" json:"-"`
 	CreatedAt    time.Time  `json:"created_at"`
